#521 merge jms-1 and jms-2 into jms
