+++ opentracing-specialagent/pom.xml
@@ -157,14 +157,7 @@
         </dependency>
         <dependency>
           <groupId>io.opentracing.contrib.specialagent.rule</groupId>
-          <artifactId>jms-1</artifactId>
-          <version>${project.version}</version>
-          <optional>true</optional>
-          <scope>provided</scope>
-        </dependency>
-        <dependency>
-          <groupId>io.opentracing.contrib.specialagent.rule</groupId>
-          <artifactId>jms-2</artifactId>
+          <artifactId>jms</artifactId>
           <version>${project.version}</version>
           <optional>true</optional>
           <scope>provided</scope>
