--- rule/jms-1/README.md
@@ -1,11 +1,0 @@
-# SpecialAgent Rule for JMS v1
-
-**Rule Name:** `jms:1`
-
-## Compatibility
-
-```xml
-<groupId>javax.jms</groupId>
-<artifactId>jms-api</artifactId>
-<version>[1.1-rev-1,LATEST]</version>
-```
