--- rule/jms-1/src/test/java/io/opentracing/contrib/specialagent/rule/jms1/Jms1Test.java
@@ -1,50 +1,0 @@
-/* Copyright 2019 The OpenTracing Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.opentracing.contrib.specialagent.rule.jms1;
-
-import javax.jms.JMSException;
-import javax.jms.Session;
-
-import org.apache.activemq.ActiveMQConnectionFactory;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.runner.RunWith;
-
-import io.opentracing.contrib.specialagent.AgentRunner;
-import io.opentracing.mock.MockTracer;
-
-@RunWith(AgentRunner.class)
-public class Jms1Test extends JmsTest {
-  @Before
-  public void before(final MockTracer tracer) {
-    tracer.reset();
-  }
-
-  @BeforeClass
-  public static void startActiveMQ() throws JMSException {
-    final ActiveMQConnectionFactory connectionFactory = new ActiveMQConnectionFactory("vm://localhost?broker.persistent=false");
-    connection = connectionFactory.createConnection();
-    connection.start();
-    session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE);
-  }
-
-  @AfterClass
-  public static void stopActiveMQ() throws JMSException {
-    session.close();
-    connection.close();
-  }
-}
