--- rule/jms-2/README.md
@@ -1,11 +1,0 @@
-# SpecialAgent Rule for JMS v2
-
-**Rule Name:** `jms:2`
-
-## Compatibility
-
-```xml
-<groupId>javax.jms</groupId>
-<artifactId>javax.jms-api</artifactId>
-<version>[2.0.1,LATEST]</version>
-```
