--- rule/jms-2/pom.xml
@@ -1,66 +1,0 @@
-<!--
-  Copyright 2019 The OpenTracing Authors
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<project
-  xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>io.opentracing.contrib.specialagent.rule</groupId>
-    <artifactId>rule</artifactId>
-    <version>1.6.2-SNAPSHOT</version>
-  </parent>
-  <artifactId>jms-2</artifactId>
-  <name>SpecialAgent Rule for JMS v2</name>
-  <properties>
-    <sa.rule.name>jms:2</sa.rule.name>
-    <min.version>2.0.1</min.version>
-    <passCompatibility>javax.jms:javax.jms-api:[${min.version},]</passCompatibility>
-    <failCompatibility>javax.jms:jms-api:[1.1-rev-1,]</failCompatibility>
-    <version.artemis>2.7.0</version.artemis>
-    <version.opentracing.jms>0.1.7</version.opentracing.jms>
-  </properties>
-  <dependencies>
-    <dependency>
-      <!-- https://github.com/opentracing-contrib/java-jms -->
-      <groupId>io.opentracing.contrib</groupId>
-      <artifactId>opentracing-jms-2</artifactId>
-      <version>${version.opentracing.jms}</version>
-      <optional>true</optional>
-    </dependency>
-    <dependency>
-      <groupId>javax.jms</groupId>
-      <artifactId>javax.jms-api</artifactId>
-      <version>${min.version}</version>
-      <optional>true</optional>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.activemq</groupId>
-      <artifactId>artemis-server</artifactId>
-      <version>${version.artemis}</version>
-      <optional>true</optional>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.activemq</groupId>
-      <artifactId>artemis-jms-client</artifactId>
-      <version>${version.artemis}</version>
-      <optional>true</optional>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-</project>
