+++ rule/jms/README.md
@@ -1,0 +1,21 @@
+# SpecialAgent Rule for JMS v1 and v2
+
+**Rule Name:** `jms`
+
+## Compatibility
+
+### JMS v1
+
+```xml
+<groupId>javax.jms</groupId>
+<artifactId>jms-api</artifactId>
+<version>[1.1-rev-1,LATEST]</version>
+```
+
+### JMS v2
+
+```xml
+<groupId>javax.jms</groupId>
+<artifactId>javax.jms-api</artifactId>
+<version>[2.0.1,LATEST]</version>
+```
