+++ rule/pom.xml
@@ -51,8 +51,7 @@
         <module>aws-sdk-2</module>
         <module>cassandra-driver-3</module>
         <module>cassandra-driver-4</module>
-        <module>jms-1</module>
-        <module>jms-2</module>
+        <module>jms</module>
         <module>elasticsearch-client-rest</module>
         <module>elasticsearch-client-transport</module>
         <module>rxjava-2</module>
